发布（全流程）+批改（全流程）

+++ src/main/java/com/lsm/frame/mapper/CourseJobUserMapper.java
@@ -14,7 +14,7 @@ public interface CourseJobUserMapper {
 
     int insertSelective(CourseJobUser record);
 
-    CourseJobUser selectByPrimaryKey(Integer id);
+    CourseJobUser selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(CourseJobUser record);
